requirements.txt
click==7.1.2
invoke==1.5.0
pymongo==3.11.3
python-bsonjs==0.2.0
